[GTK] Initialize text-antialias when initializing GC object
On some platforms, the default text-antialias of the system might differ
from Cairo. When creating a new GC object is created, the Cairo default
is used. But when setTextAntialias(SWT.DEFAULT) is called, the system
default is used instead.

To avoid this ambiguity, setTextAntialias(SWT.DEFAULT) should always be
called when creating a GC object, in order to always use the system
default.