[GTK] Initialize text-antialias when initializing Image GC object
If a string is drawn without ever calling setTextAntialias(), the Cairo
default is used. If this method is called with SWT.DEFAULT, the system
default is used. Because those values are not guaranteed to be equal,
this may then lead to different results.

To avoid this ambiguity, the GC needs to be explicitly initialized with
the system default value.