Use accessibility cursor size when using provider-based cursor
With 6e4241da7ffbe82ca6e307a794bcb8b5be661a00, the cursor size set in
the Windows setting is included in the total zoom factor of the cursor.
However, this is only the case when the cursor is created using an
ImageData object.

This change corrects this oversight and also uses this factor when
creating a cursor using an ImageDataProvider.